+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -132,7 +132,7 @@ namespace UnityEngine.InputSystem
                 ////TODO: if we're *inside* an update, this should use the current update type
 
                 #if UNITY_EDITOR
-                if (!gameIsPlayingAndHasFocus)
+                if (!gameIsPlaying || !hasInputFocus)
                     return InputUpdateType.Editor;
                 #endif
 
@@ -296,9 +296,34 @@ namespace UnityEngine.InputSystem
             }
         }
 
-        private bool gameIsPlayingAndHasFocus =>
 #if UNITY_EDITOR
-                     m_Runtime.isInPlayMode && !m_Runtime.isPaused && (m_HasFocus || InputEditorUserSettings.lockInputToGameView);
+        private bool m_RunUpdatesInEditMode;
+
+        public bool runUpdatesInEditMode
+        {
+            get => m_RunUpdatesInEditMode;
+            set => m_RunUpdatesInEditMode = value;
+        }
+#endif
+
+        private bool gameIsPlaying =>
+#if UNITY_EDITOR
+                     m_Runtime.isInPlayMode && !m_Runtime.isPaused;
+#else
+            true;
+#endif
+
+        // TODO Should we assume focus is always true outside the editor?
+        private bool hasInputFocus =>
+#if UNITY_EDITOR
+            (m_HasFocus || InputEditorUserSettings.lockInputToGameView);
+#else
+            true;
+#endif
+
+        private bool shouldProcessInputEvents =>
+#if UNITY_EDITOR
+            (gameIsPlaying || runUpdatesInEditMode) && hasInputFocus;
 #else
             true;
 #endif
@@ -2531,10 +2556,10 @@ namespace UnityEngine.InputSystem
             var mask = m_UpdateMask;
 #if UNITY_EDITOR
             // Ignore editor updates when the game is playing and has focus. All input goes to player.
-            if (gameIsPlayingAndHasFocus)
+            if (gameIsPlaying && hasInputFocus)
                 mask &= ~InputUpdateType.Editor;
             // If the player isn't running, the only thing we run is editor updates.
-            else if (updateType != InputUpdateType.Editor)
+            else if (updateType != InputUpdateType.Editor && !runUpdatesInEditMode)
                 return false;
 #endif
             return (updateType & mask) != 0;
@@ -2613,14 +2638,14 @@ namespace UnityEngine.InputSystem
             //       in the buffer and having older timestamps will get rejected.
 
             var currentTime = updateType == InputUpdateType.Fixed ? m_Runtime.currentTimeForFixedUpdate : m_Runtime.currentTime;
-            var timesliceEvents = gameIsPlayingAndHasFocus && InputSystem.settings.updateMode == InputSettings.UpdateMode.ProcessEventsInFixedUpdate;
+            var timesliceEvents = shouldProcessInputEvents && InputSystem.settings.updateMode == InputSettings.UpdateMode.ProcessEventsInFixedUpdate;
 
             // Early out if there's no events to process.
             if (eventBuffer.eventCount <= 0)
             {
                 // Normally, we process action timeouts after first processing all events. If we have no
                 // events, we still need to check timeouts.
-                if (gameIsPlayingAndHasFocus)
+                if (shouldProcessInputEvents)
                     ProcessStateChangeMonitorTimeouts();
 
                 #if ENABLE_PROFILER
@@ -2801,8 +2826,13 @@ namespace UnityEngine.InputSystem
 
                         // Update timestamp on device.
                         // NOTE: We do this here and not in UpdateState() so that InputState.Change() will *NOT* change timestamps.
-                        //       Only events should.
-                        if (device.m_LastUpdateTimeInternal <= eventPtr.internalTime)
+                        //       Only events should. If running play mode updates in editor, we want to defer to the play mode
+                        //       callbacks to set the last update time to avoid dropping events only processed by the editor state.
+                        if (device.m_LastUpdateTimeInternal <= eventPtr.internalTime
+#if UNITY_EDITOR
+                            && !(updateType == InputUpdateType.Editor && runUpdatesInEditMode)
+#endif
+                        )
                             device.m_LastUpdateTimeInternal = eventPtr.internalTime;
 
                         // Make device current. Again, only do this when receiving events.
@@ -2867,7 +2897,17 @@ namespace UnityEngine.InputSystem
                         break;
                 }
 
-                m_InputEventStream.Advance(leaveEventInBuffer: false);
+                // Editor updates go into a separate buffer.  If we want to update the player buffer,
+                // we need to keep the events in the queue, and reprocess again once we are in a player-based update loop.
+#if UNITY_EDITOR
+                bool leaveInBuffer = updateType == InputUpdateType.Editor &&
+                    runUpdatesInEditMode &&
+                    (currentEventReadPtr->type == StateEvent.Type ||
+                        currentEventReadPtr->type == DeltaStateEvent.Type);
+#else
+                bool leaveInBuffer = false;
+#endif
+                m_InputEventStream.Advance(leaveEventInBuffer: leaveInBuffer);
             }
 
             m_Metrics.totalEventProcessingTime += ((double)(Stopwatch.GetTimestamp() - processingStartTime)) / Stopwatch.Frequency;
@@ -2880,7 +2920,7 @@ namespace UnityEngine.InputSystem
 
             m_InputEventStream.Close(ref eventBuffer);
 
-            if (gameIsPlayingAndHasFocus)
+            if (shouldProcessInputEvents)
                 ProcessStateChangeMonitorTimeouts();
 
             ////TODO: fire event that allows code to update state *from* state we just updated
@@ -3225,7 +3265,7 @@ namespace UnityEngine.InputSystem
             // Updates go to the editor only if the game isn't playing or does not have focus.
             // Otherwise we fall through to the logic that flips for the *next* dynamic and
             // fixed updates.
-            if (updateType == InputUpdateType.Editor && !gameIsPlayingAndHasFocus)
+            if (updateType == InputUpdateType.Editor && (!gameIsPlaying || !hasInputFocus))
             {
                 ////REVIEW: This isn't right. The editor does have update ticks which constitute the equivalent of player frames.
                 // The editor doesn't really have a concept of frame-to-frame operation the